feat!: Don't bundle any syntax highlighting themes
Also load in languages dynamically. This only happens on the server, all
languages will be bundled on the client. However, only GitHub Dark will
be bundled with Znak, any other themes can be added by adding Shiki as a
dependency and importing them directly. This also allows users to make
and add in their own themes using the CodeTheme type.