feat: Switch KaTeX for Temml
There is no difference in output as it is still MathML, so this isn't a
breaking change. However, this change means that the browser bundle is
slightly smaller as Temml is a smaller library than KaTeX.

Future exploration: Extract only the renderToString part of the Temml
library and make a separate library specifically for this purpose. This
will remove all of the code needed for the `render()` browser functions,
making it significantly lighter.